+++ website/src/examples/mod.rs
@@ -1,14 +1,16 @@
 use leptos::*;
-use wasm_bindgen::{JsValue, JsCast};
+use wasm_bindgen::{JsCast, JsValue};
 use wasm_bindgen_futures::JsFuture;
-use web_sys::{Request, Response, RequestInit};
+use web_sys::{Request, RequestInit, Response};
 
 use crate::components::CodeBlock;
 
+mod basic_box;
 mod basic_dropdown;
 mod basic_modal;
 mod general_form;
 
+pub use basic_box::BasicBox;
 pub use basic_dropdown::BasicDropdown;
 pub use basic_modal::BasicModal;
 pub use general_form::GeneralForm;
@@ -17,8 +19,13 @@ async fn get_code_example(name: &'static str) -> Result<String, JsValue> {
     let opts = RequestInit::new();
     let code_example_path = format!("/examples/{name}.rs");
     let request = Request::new_with_str_and_init(&code_example_path, &opts)?;
-    let response = JsFuture::from(window().fetch_with_request(&request)).await?.dyn_into::<Response>()?;
-    JsFuture::from(response.text()?).await?.as_string().ok_or(JsValue::UNDEFINED)
+    let response = JsFuture::from(window().fetch_with_request(&request))
+        .await?
+        .dyn_into::<Response>()?;
+    JsFuture::from(response.text()?)
+        .await?
+        .as_string()
+        .ok_or(JsValue::UNDEFINED)
 }
 
 #[component]