+++ website/src/pages/mod.rs
@@ -1,10 +1,12 @@
 mod components_page;
+mod elements_page;
 mod form_page;
 mod guides_page;
 mod home_page;
 mod not_found_page;
 
 pub use components_page::ComponentsPage;
+pub use elements_page::ElementsPage;
 pub use form_page::FormPage;
 pub use guides_page::GuidesPage;
 pub use home_page::HomePage;